+++ app/src/main/java/hr/ferit/measureup/sensing/sensors/Thermometer.kt
@@ -1,8 +1,9 @@
-package hr.ferit.measureup.sensors
+package hr.ferit.measureup.sensing.sensors
 
 import android.content.Context
 import android.content.pm.PackageManager
 import android.hardware.Sensor
+import hr.ferit.measureup.sensing.AndroidSensor
 
 class Thermometer(context: Context) :
     AndroidSensor(
