+++ app/src/main/java/hr/ferit/measureup/ui/viewmodels/LightSensorViewModel.kt
@@ -1,9 +1,7 @@
 package hr.ferit.measureup.ui.viewmodels
 
-import androidx.compose.runtime.getValue
 import androidx.lifecycle.ViewModel
-import hr.ferit.measureup.sensors.LightSensor
-import hr.ferit.measureup.sensors.MeasurableSensor
+import hr.ferit.measureup.sensing.MeasurableSensor
 import kotlinx.coroutines.flow.MutableStateFlow
 
 class LightSensorViewModel(private val lightSensor: MeasurableSensor) : ViewModel() {
